requirements.txt
altair==5.5.0
antlr4-python3-runtime==4.9.3
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.13.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
iopath==0.1.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.24.1
networkx==3.3
numpy==2.1.2
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==3.1.1
protobuf==5.29.3
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
streamlit==1.40.0
streamlit-drawable-canvas==0.9.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
toml==0.10.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==5.0.3
yacs==0.1.8
altair==5.5.0
antlr4-python3-runtime==4.9.3
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.13.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
iopath==0.1.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.24.1
networkx==3.3
numpy==2.1.2
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==3.1.1
protobuf==5.29.3
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
streamlit==1.40.0
streamlit-drawable-canvas @ git+https://github.com/VikParuchuri/streamlit-drawable-canvas@c6b9290ddfae8f347c652986cfd844065c6195c5
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
toml==0.10.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==5.0.3
yacs==0.1.8
altair==5.5.0
antlr4-python3-runtime==4.9.3
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.13.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
iopath==0.1.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.24.1
networkx==3.3
numpy==2.1.2
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==3.1.1
protobuf==5.29.3
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
streamlit==1.40.0
streamlit-drawable-canvas @ git+https://github.com/VikParuchuri/streamlit-drawable-canvas@c6b9290ddfae8f347c652986cfd844065c6195c5
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
toml==0.10.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==5.0.3
yacs==0.1.8
